+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/internal/video/DefaultVideoClientController.kt
@@ -34,6 +34,8 @@ import com.xodee.client.video.VideoSubscriptionConfigurationInternal
 import java.security.InvalidParameterException
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.sync.Mutex
+import kotlinx.coroutines.sync.withLock
 
 class DefaultVideoClientController(
     private val context: Context,
@@ -64,7 +66,7 @@ class DefaultVideoClientController(
     private val cameraCaptureSource: DefaultCameraCaptureSource
     private var videoSourceAdapter = VideoSourceAdapter()
     private var isUsingInternalCaptureSource = false
-
+    private val videoClientStopMutex = Mutex()
     init {
         videoClientStateController.bindLifecycleHandler(this)
 
@@ -102,10 +104,16 @@ class DefaultVideoClientController(
             // So it doesn't call it spuriously
             videoClientObserver.primaryMeetingPromotionObserver = null
 
-            videoClientStateController.stop()
+            videoClientStopMutex.withLock {
+                // Race conditions in video client stop functions may lead to crashes at pointer deallocation
+                // if multiple such threads are called at the same time.
+                videoClientStateController.stop()
 
-            eglCore?.release()
-            eglCore = null
+                // We put release eglCore under mutex as a race condition may decrease its reference counter
+                // below 0 and cause crashes if multiple such calls happen together.
+                eglCore?.release()
+                eglCore = null
+            }
         }
     }